requirements.txt
numpy>=1.21.6
Pillow==10.3.0
pyautogui==0.9.54
pyperclip==1.8.2
pytesseract==0.3.10
pytest==7.2.2
python_Levenshtein==0.20.9
colorama==0.4.6
opencv-python==4.10.0.84
mss==9.0.1
qat==1.1.2
pyyaml==5.3.1
pyscreeze==0.1.21